fuzzgen: Move `Testcase` to fuzzgen fuzzer
