+++ fuzz/Cargo.toml
@@ -16,6 +16,7 @@ cranelift-wasm = { workspace = true }
 cranelift-filetests = { workspace = true }
 cranelift-interpreter = { workspace = true }
 cranelift-fuzzgen = { workspace = true }
+cranelift-native = { workspace = true }
 libfuzzer-sys = { version = "0.4.0", features = ["arbitrary-derive"] }
 target-lexicon = { workspace = true }
 smallvec = { workspace = true }
